+++ lib/specs/giphy_dialog_method_channel.dart
@@ -74,4 +74,3 @@ class GiphyDialogMethodChannel extends GiphyDialogPlatform {
     await _channel.invokeMethod('hide');
   }
 }
-