+++ lib/specs/giphy_dialog_platform_interface.dart
@@ -97,4 +97,3 @@ abstract class GiphyDialogPlatform extends PlatformInterface {
     throw UnimplementedError('hide() has not been implemented.');
   }
 }
-