+++ lib/specs/giphy_flutter_sdk_method_channel.dart
@@ -30,4 +30,3 @@ class GiphyFlutterSdkPlatformChannel extends GiphyFlutterSdkPlatform {
     });
   }
 }
-