+++ lib/specs/giphy_flutter_sdk_platform_interface.dart
@@ -51,4 +51,3 @@ abstract class GiphyFlutterSdkPlatform extends PlatformInterface {
     throw UnimplementedError('configure() has not been implemented.');
   }
 }
-