+++ lib/specs/giphy_video_manager_method_channel.dart
@@ -39,4 +39,3 @@ class GiphyVideoManagerMethodChannel extends GiphyVideoManagerPlatform {
     await _channel.invokeMethod('resume');
   }
 }
-